p1 - 03.08.09 - managing products and wishlists exercise (solution) - update code to match step 2 of the course

+++ src/main/java/com/springbootfundamentals/store/repositories/CategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.springbootfundamentals.store.entities.Category;
 
-public interface CategoryRepository extends CrudRepository<Category, Long> {
+public interface CategoryRepository extends CrudRepository<Category, Byte> {
     
 }
